Fait en sorte que emprunter fonctionne

+++ src/main/java/fr/dta/adherent/controller/AdherentController.java
@@ -21,6 +21,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import fr.dta.adherent.dao.AdherentDAO;
 import fr.dta.adherent.modele.Adherent;
+import fr.dta.configuration.View;
 import fr.dta.media.modele.Media;
 
 @RestController
@@ -31,7 +32,7 @@ public class AdherentController {
 	AdherentDAO dao;
 	
 	@RequestMapping (value = "{id}", method = RequestMethod.GET)
-	public Adherent rechercheAdherent( @PathVariable Long id) {
+	public Adherent rechercheAdherent( @PathVariable Integer id) {
 		return dao.findOne(id);
 	}
